+++ VoiceUDPRecive/VoiceUDPSender.cpp
@@ -4,7 +4,7 @@ void VoiceUDPSender::StartUp()
 {
 	WSACleanup();
 	WSAStartup(MAKEWORD(2, 0), &wsaData);   //MAKEWORD(2, 0)はwinsockのバージョン2.0ってこと
-	sock = socket(AF_INET, SOCK_DGRAM, 0);  //AF_INETはIPv4、SOCK_DGRAMはUDP通信、0は？
+	sock = static_cast<int>(socket(AF_INET, SOCK_DGRAM, 0));  //AF_INETはIPv4、SOCK_DGRAMはUDP通信、0は？
 	addr.sin_family = AF_INET;
 	addr.sin_port = htons(50008);// 待ち受けポート番号を50008にする
 	addr.sin_addr.S_un.S_addr = inet_addr("127.0.0.1");// 送信アドレスを設定
